Add buttons for jumping to next/previsous bar
Also keep track of the bar we are in and, after stopping, continue
playback from said bar.

The old waitForPlay mechanism seemed to get stuck when events were not
taking place in the expected order. It sometimes happened that when
skipping bars, the player thread arrived at waitForPlay() only after
the UI thread had already called play() leading the player thread to
wait indefinitely. Replace this with a mechanism where the player
thread does not need to wait if it already has permission to play.