Bugfix: Fix measure calculation that got broken in 1db5a56
In said refactoring, I forgot that the first event of a measure is not
necessarily at the measure start. Make start tick a property of Measure
and calculate it when parsing Measures.

Note that this also fixes click injection in cases where there isn't
any event at measure start. Also give the test a track without the
click so as to make the test more focused (a bit clumsy but works for
now).